AppBuilder9000/requirements.txt
django==2.2.5
djangorestframework==3.11.0
django-bootstrap4==1.1.1
django-crispy-forms==1.7.2
beautifulsoup4==4.8.0
requests==2.22.0